+++ CHANGELOG.md
@@ -4,15 +4,37 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
+
 ## [4.x.x]
+
+### Added
+
+
+### Fixed
+
+
+
+## [4.4.1]
+
+### Added
+- Display SV rank model version when available
+
+### Fixed
+- Fixed upload of delivery report via API
+
+
+## [4.4.0]
 
 ### Added
 - Displaying more info on the Causatives page and hiding those not causative at the case level
 - Add a comment text field to Sanger order request form, allowing a message to be included in the email
 - MatchMaker Exchange integration
+- List cases with empty synopsis, missing HPO terms and phenotype groups.
+- Search for cases with open research list, or a given case status (active, inactive, archived)
 
 ### Fixed
 - Variant query builder split into several functions
+- Fixed delivery report load bug
 
 
 ## [4.3.3]